Se extrae obra social del formulario
